LIBAVALON-398. Re-implement Master File downloads for Avalon 7.8
As discussed in the July 24, 2025 Digital Collections Daily Scrum, the
"Downloads" button and "Downloads" panel are being replaced in
Avalon 7.8 by customizing the "Files" tab provided by the
"MediaObjectRamp" React component to display the master files for
download, as well as the supplemental files.

A “app/javascript/components/UmdMasterFiles.jsx” React component (named
to be consistent with the samvera/ramp “SupplementalFiles.jsx”
component) was created to support adding the master file download links
to the “Files” tab in the “MediaObjectRamp” React component.

A “master_files_downloads” method was added to
“app/controllers/media_objects_controller.rb” to generate the data
needed (id, filename, file label, and url) for creating the download
links. This information is passed to the React component. The
“app/views/media_objects/_item_view.html.erb” was modified to pass the
master file information (as well as whether the user had permission to
download the master files) to the “MediaObjectRamp” React component.

The Avalon-provided “MediaObjectRamp” component in the
“components/MediaObjectRamp.jsx” file was modified to incorporate the
“UmdMasterFile” React component into the “Files” tab, along with logic
around when the information should be displayed.

The “/app/views/media_objects/_download.html.erb” was deleted, as well
as UMD customizations in “app/assets/stylesheets/avalon/_buttons.scss”,
as they are no longer needed.

Updated the “umd_docs/UmdCustomizations.md” with information about the
“Master File Downloads” customization.

https://umd-dit.atlassian.net/browse/LIBAVALON-398